add validation to category

+++ MS_Finanzas/src/main/java/com/example/demo/infra/persistence/repository/JpaCategoryRepositoryAdapter.java
@@ -7,6 +7,8 @@ import java.util.UUID;
 import org.springframework.stereotype.Component;
 
 import com.example.demo.application.repository.CategoryRepositoryPort;
+import com.example.demo.domain.exception.CategoryAlreadyExistsException;
+import com.example.demo.domain.exception.CategoryInUseException;
 import com.example.demo.domain.exception.ResourceNotFoundException;
 import com.example.demo.domain.model.Category;
 import com.example.demo.infra.mapper.CategoryEntityMapper;
@@ -15,10 +17,12 @@ import com.example.demo.infra.persistence.entity.CategoryEntity;
 @Component
 public class JpaCategoryRepositoryAdapter implements CategoryRepositoryPort {
     private final JpaCategoryRepository jpaCategoryRepository;
+    private final JpaTransactionRepository jpaTransactionRepository;
     private final CategoryEntityMapper categoryEntityMapper;
 
-    public JpaCategoryRepositoryAdapter(JpaCategoryRepository jpaCategoryRepository, CategoryEntityMapper categoryEntityMapper) {
+    public JpaCategoryRepositoryAdapter(JpaCategoryRepository jpaCategoryRepository, JpaTransactionRepository jpaTransactionRepository, CategoryEntityMapper categoryEntityMapper) {
         this.jpaCategoryRepository = jpaCategoryRepository;
+        this.jpaTransactionRepository = jpaTransactionRepository;
         this.categoryEntityMapper = categoryEntityMapper;
     }
 
@@ -35,6 +39,10 @@ public class JpaCategoryRepositoryAdapter implements CategoryRepositoryPort {
 
     @Override
     public Category save(Category category) {
+        if (jpaCategoryRepository.existsByNombre(category.nombre())) {
+            throw new CategoryAlreadyExistsException(category.nombre());
+        }
+
         CategoryEntity categoryEntity = categoryEntityMapper.toEntity(category);
         CategoryEntity savedCategoryEntity = jpaCategoryRepository.save(categoryEntity);
         return categoryEntityMapper.toDomain(savedCategoryEntity);
@@ -42,6 +50,10 @@ public class JpaCategoryRepositoryAdapter implements CategoryRepositoryPort {
 
     @Override
     public Category update(UUID categoryId, Category category) {
+        if (jpaCategoryRepository.existsByNombre(category.nombre())) {
+            throw new CategoryAlreadyExistsException(category.nombre());
+        }
+
         CategoryEntity existingCategoryEntity = jpaCategoryRepository.findById(categoryId)
         .orElseThrow(() -> new ResourceNotFoundException("La categoria no fue encontrada"));
         existingCategoryEntity.setNombre(category.nombre());
@@ -55,6 +67,16 @@ public class JpaCategoryRepositoryAdapter implements CategoryRepositoryPort {
         if (!jpaCategoryRepository.existsById(categoryId)) {
             throw new ResourceNotFoundException("La categoria no fue encontrada");
         }
+
+        if (jpaTransactionRepository.existsByCategoryId(categoryId)) {
+            throw new CategoryInUseException(categoryId.toString());
+        }
+
         jpaCategoryRepository.deleteById(categoryId);
+    }
+
+    @Override
+    public boolean existsByName(String name) {
+        return jpaCategoryRepository.existsByNombre(name);
     }
 }
